fix: create admin_policy_acceptance table with unique constraint
The table was defined in schema_policy.ts which was excluded from
drizzle.config.ts, so it was never created in the DB. The missing
table caused the POST /api/admin/policy/accept route to fail, which
meant setPolicyAccepted(true) was never called after clicking agree,
leaving the user stuck on the acceptance modal.

Fixes:
- Move adminPolicyAcceptance table definition into schema.ts so
  drizzle-kit push creates it
- Add a unique index on (adminId, policyVersion, termsVersion,
  privacyVersion) required for onConflictDoUpdate to work correctly
- Reduce schema_policy.ts to a re-export shim to keep existing
  imports working
- Add migration 0003 for the new table and unique index
